refactor: reorganize codeagent backend configuration and update tests
