+++ src/vibe3/agents/backends/codeagent.py
@@ -1,4 +1,3 @@
-import json
 import re
 import shlex
 import subprocess
@@ -6,10 +5,12 @@ import sys
 import tempfile
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Final, cast
+from typing import Final, cast
 
-from loguru import logger
-
+from vibe3.agents.backends.codeagent_config import (
+    resolve_effective_agent_options,
+    sync_models_json,
+)
 from vibe3.exceptions import AgentExecutionError
 from vibe3.models.review_runner import AgentOptions, AgentResult
 
@@ -18,11 +19,6 @@ DEFAULT_WRAPPER_PATH: Final[Path] = (
     Path.home() / ".claude" / "bin" / "codeagent-wrapper"
 )
 
-# Path to codeagent models config
-MODELS_JSON_PATH: Final[Path] = Path.home() / ".codeagent" / "models.json"
-REPO_MODELS_JSON_PATH: Final[Path] = (
-    Path(__file__).resolve().parents[4] / "config" / "models.json"
-)
 KNOWN_CODEX_STATE_DB_WARNINGS: Final[tuple[str, ...]] = (
     r"failed to open state db at .*migration .*missing in the resolved migrations",
     r"failed to initialize state runtime at .*migration "
@@ -46,112 +42,6 @@ class AsyncExecutionHandle:
     tmux_session: str
     log_path: Path
     prompt_file_path: Path
-
-
-def _read_models_json(path: Path) -> dict[str, Any]:
-    """Read a models.json file and return a dict, or empty dict on failure."""
-    try:
-        if path.exists():
-            data = json.loads(path.read_text())
-            if isinstance(data, dict):
-                return data
-    except Exception as exc:
-        logger.bind(domain="review_runner", path=str(path)).warning(
-            f"Failed to read models config: {exc}"
-        )
-    return {}
-
-
-def resolve_repo_agent_preset(
-    agent_name: str,
-) -> tuple[str | None, str | None] | None:
-    """Resolve agent preset from repo-local config/models.json.
-
-    Returns:
-        (backend, model) when repo-local mapping exists, otherwise None.
-    """
-    data = _read_models_json(REPO_MODELS_JSON_PATH)
-    agents = data.get("agents")
-    if not isinstance(agents, dict):
-        return None
-    raw = agents.get(agent_name)
-    if not isinstance(raw, dict):
-        return None
-    backend = raw.get("backend")
-    model = raw.get("model")
-    if backend is not None and not isinstance(backend, str):
-        backend = None
-    if model is not None and not isinstance(model, str):
-        model = None
-    if backend is None and model is None:
-        return None
-    return backend, model
-
-
-def resolve_effective_agent_options(options: AgentOptions) -> AgentOptions:
-    """Resolve repo-local agent preset mapping into explicit backend/model.
-
-    Priority:
-    1. Explicit backend/model override in options
-    2. Repo-local config/models.json mapping for agent preset
-    3. Fallback to raw agent mode
-    """
-    if options.backend:
-        return options
-    if not options.agent:
-        return options
-    resolved = resolve_repo_agent_preset(options.agent)
-    if not resolved:
-        return options
-    backend, mapped_model = resolved
-    return AgentOptions(
-        agent=None,
-        backend=backend,
-        model=options.model or mapped_model,
-        worktree=options.worktree,
-        timeout_seconds=options.timeout_seconds,
-    )
-
-
-def sync_models_json(options: AgentOptions) -> None:
-    """Sync effective backend/model to ~/.codeagent/models.json.
-
-    In backend mode: updates default_backend (and default_model if specified),
-    so codeagent-wrapper uses vibe3's config instead of whatever is in the file.
-
-    In agent preset mode: no-op — codeagent manages the preset's backend/model
-    from its own config.
-    """
-    effective = resolve_effective_agent_options(options)
-    if not effective.backend:
-        return  # agent preset mode — codeagent reads preset config itself
-
-    try:
-        existing: dict[str, Any] = {}
-        if MODELS_JSON_PATH.exists():
-            existing = json.loads(MODELS_JSON_PATH.read_text())
-    except Exception as exc:
-        logger.bind(domain="review_runner").warning(
-            f"Failed to read models.json, will overwrite: {exc}"
-        )
-        existing = {}
-
-    existing["default_backend"] = effective.backend
-    if effective.model:
-        existing["default_model"] = effective.model
-
-    try:
-        MODELS_JSON_PATH.parent.mkdir(parents=True, exist_ok=True)
-        MODELS_JSON_PATH.write_text(json.dumps(existing, indent=2))
-        logger.bind(
-            domain="review_runner",
-            backend=effective.backend,
-            model=effective.model,
-        ).debug("Synced models.json")
-    except Exception as exc:
-        logger.bind(domain="review_runner").warning(
-            f"Failed to write models.json: {exc}"
-        )
 
 
 def extract_session_id(stdout: str) -> str | None:
